b/e. Add check if own or permissions onto DeleteItem mutation in resolvers Mutation

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -18,7 +18,7 @@ const Mutations = {
     const item = await ctx.db.mutation.createItem(
       {
         data: {
-          // relationship betweeen the Item and the User
+          // relationship between the Item and the User
           user: {
             connect: {
               id: user.id,
@@ -50,11 +50,19 @@ const Mutations = {
     );
   },
   async deleteItem(parent, args, ctx, info) {
+    // throw new Error('Dont');
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id, title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
+    // console.log('deleteItem => item', item.user.id);
     // 2. check if user is logged in and own that item or have the permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermission = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if (!ownsItem && !hasPermission) {
+      throw new Error("You don't have permissions to do that!");
+    }
     // 3. delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -172,7 +180,7 @@ const Mutations = {
         resetTokenExpiry: null,
       },
     });
-    console.log('TCL: resetPassword -> updateUser', updateUser);
+    // console.log('TCL: resetPassword -> updateUser', updateUser);
     // Generate JWT token
     const token = jwt.sign({ userId: updateUser.id }, process.env.APP_SECRET);
     // Set JWT token onto cookie
